refactor:refactored routes and controller files
