+++ app/Http/Controllers/Backend/CacheController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Backend\Modules\Settings;
+namespace App\Http\Controllers\Backend;
 
 use App\Http\Controllers\Controller;
