+++ app/Http/Controllers/Backend/SettingsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Controllers\Backend\Modules\Settings;
+namespace App\Http\Controllers\Backend;
 
 use App\Helpers\UploadHelper;
 use App\Http\Controllers\Controller;
